GpuDataController as a pointer in UniverseImpl

+++ SimulationEngine/main.cu
@@ -18,7 +18,7 @@ int main(int argc, char *argv[])
 
 	unsigned long particleCount = 50;
 	unsigned int frameRate = 60;
-	double seconds = 10;
+	unsigned int seconds = 10;
 	double meanMass = 0.01;	
 	double starMass = 50;
 	double meanSpeed = 0.04;
